Switched to a js shim.
